Refactor: move optional params function from gmws.go to parameters.go
