+++ gmws/basic_structs.go
