+++ gmws/parameters.go
@@ -16,7 +16,7 @@ func formatParameterKey(baseKey string, keys ...string) string {
 	return strings.Join(paramKeys, ".")
 }
 
-// Parameters is the parameters pass to the gomws api.
+// Parameters is the parameters pass to the gomws API.
 type Parameters map[string]interface{}
 
 // Merge merge the target Parameters to current Parameters.
@@ -36,7 +36,7 @@ func (params Parameters) Merge(parameters Parameters) Parameters {
 // 	p := Parameters{
 // 		"slice": []string{"a", "b"},
 // 	}
-// 	p.StructureKeys("arrayFiled", "fields")
+// 	p.StructureKeys("slice", "fields")
 // result:
 // 	Parameters{
 // 		"slice.fields.1": "a",
@@ -116,4 +116,42 @@ func (params Parameters) Normalize() (mwsHttps.Values, error) {
 		nParams.Set(key, stringVal)
 	}
 	return nParams, nil
+}
+
+// OptionalParams get the values from the pass in parameters.
+// Only values for keys that are accepted will be returned.
+//
+// Note: The keys returned will be in title case.
+//
+// If the key appear in mulit parameters, later one will override the previous.
+// Ex:
+// 		ps := []mwsHttps.Parameters{
+// 			{"key1": "value1", "key2": "value2"},
+// 			{"key1": "newValue1", "key3": "value3"},
+// 		}
+// 		acceptKeys := []string{"key1", "key2"}
+// 		resultParams := OptionalParams(acceptKeys, ps)
+// result:
+// 		resultParams -> {"Key1": "newValue1", "Key2": "value2"}
+func OptionalParams(acceptKeys []string, ops []Parameters) Parameters {
+	param := Parameters{}
+	op := Parameters{}
+
+	if len(ops) == 0 {
+		return param
+	}
+
+	for _, p := range ops {
+		op.Merge(p)
+	}
+
+	for _, key := range acceptKeys {
+		value, ok := op[key]
+		if ok {
+			param[strings.Title(key)] = value
+			delete(op, key)
+		}
+	}
+
+	return param
 }
