+++ gmws/xml_parser.go
@@ -10,18 +10,18 @@ import (
 	"github.com/svvu/gomws/mwsHttps"
 )
 
-// XMLParser use parse the xml string
+// XMLParser use parse the XML string.
 type XMLParser struct {
 	XMLString string
 }
 
-// NewXMLParser Create a new parse for the response, seting the response result to XMLString
+// NewXMLParser Create a new parse for the response, seting the response result to XMLString.
 func NewXMLParser(response *mwsHttps.Response) *XMLParser {
 	return &XMLParser{XMLString: response.Result}
 }
 
-// PrettyPrint Print the xml in indent format.
-// Note: The method will ignore namespace, attributes, comments for the tag
+// PrettyPrint Print the XML in indent format.
+// Note: The method will ignore namespace, attributes, comments for the tag.
 func (xmlp *XMLParser) PrettyPrint() {
 	decoder := xml.NewDecoder(strings.NewReader(xmlp.XMLString))
 	decoder.Strict = false
@@ -48,7 +48,7 @@ func (xmlp *XMLParser) PrettyPrint() {
 	fmt.Println(string(rBuffer.Bytes()))
 }
 
-// Parse unmarshal the xml string to target struct
+// Parse unmarshal the XML string to target struct
 func (xmlp *XMLParser) Parse(v interface{}) error {
 	err := xml.Unmarshal([]byte(xmlp.XMLString), v)
 
@@ -60,7 +60,7 @@ func (xmlp *XMLParser) Parse(v interface{}) error {
 	return err
 }
 
-// XMLResultHandler is interface to allow parsed xml result to have callback
+// XMLResultHandler is interface to allow parsed XML result to have callback
 type XMLResultHandler interface {
 	ParseCallback()
 }
