+++ mwsHttps/response.go
@@ -5,6 +5,7 @@ import (
 	"strconv"
 )
 
+// Response is the reponse from the API.
 type Response struct {
 	Result     string
 	Error      error
@@ -12,8 +13,8 @@ type Response struct {
 	Status     string
 }
 
-// CheckStatusCode check whether or not the status indicate a request error
-// When the code not start with 1 or 2, false returned
+// CheckStatusCode check whether or not the status indicate a request error.
+// When the code not start with 1 or 2, false returned.
 func CheckStatusCode(code int) bool {
 	scode := strconv.Itoa(code)
 	greenStatus := regexp.MustCompile(`^[1-2][0-9]{2}$`)
